refactor(a2a): extract TaskManager and JSON-RPC into pais/a2a.py module
Move task lifecycle orchestration (submit, execute, cancel) from AgentServer
into TaskManager class that wraps TaskStore with execution logic and asyncio
task tracking. Move JSON-RPC models, dispatcher, and method handlers from
server.py/serverutils.py into dedicated a2a.py module. server.py now calls
setup_a2a_routes() to mount the A2A endpoint. TaskStore remains pure storage.
